Rearrange and add relevant info
Made some changes/additions to the readme to reflect the other properties we now have going (ie. the newsletter, website and event listing). Also added the NodeSchool Toronto Badge and @tessalt / @vladikoff to the Mentors list (feel free to remove yourselves or reach out to me if you'd rather not be listed as mentors)